+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Display quotes table page with seller quotes if the quote was splitted
+- Display quotes children quantity and seller trade name when there are splitted quotes
 
 ## [1.6.6] - 2025-01-06
 ### Fixed